1. Changed the calc_grad_chisquare function. I now calculate using (low_indx,high_indx+1) This was found earlier, but it was found that adding the 1 needs to less smooth maps. I have now verified that without adding 1, the maps have much larger chisquare at select locations. And in any case, since I am finding values between low_indx,high_indx (inclusive of both), I should use all the values to calculate the grad_chisquare as well.
I have also added a function named remove_cluster_boundary. In remove_cluster, I was
trying to find new parameter values for the whole cluster. However at the edges of true structures,
it is possible that only part of that cluster can be given new parameter values. This new function
accepts this. This follows the same strategy as the image removal, but only difference is that
I do the minimisation in the pixel spectrum space.

I have also added a function which searches for clusters in the whole parameter map. The idea
was that since we are dealing with local parameters, it does not have information about the global
scale. However, it would be much easier to cluster in the global space. Then we can try to replace
values of the smaller clusters, starting from values of the largest cluster. However I have found
that this does not change thinsg significantly.